+++ include/ignition/math/graph/Graph.hh
@@ -18,8 +18,8 @@
 #define IGNITION_MATH_GRAPH_GRAPH_HH_
 
 #include <cassert>
-#include <iostream>
 #include <map>
+#include <ostream>
 #include <set>
 #include <string>
 #include <utility>